+++ RunCoff.sln
@@ -1,0 +1,16 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bofloader", "Bofloader\Bofloader.csproj", "{0C395721-2D3A-4DD3-9D88-124A0CF67A66}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{0C395721-2D3A-4DD3-9D88-124A0CF67A66}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0C395721-2D3A-4DD3-9D88-124A0CF67A66}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0C395721-2D3A-4DD3-9D88-124A0CF67A66}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0C395721-2D3A-4DD3-9D88-124A0CF67A66}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
